Add files via upload
have to run `npm i dotenv --save` and `npm i node-fetch --save` before using the code or it will cause an error.
the part I'm working on Is the fetch function in download.js. feel free to try fixing it. the problem is it only checks if the file is finished being transcribed once. it is supposed to keep checking until it is complete. 
transcribeAudio.js is the main file